Rename PartitionFunction to BucketPartitionFunction
This commit renames `PartitionFunction` to `BucketPartitionFunction` and
re-introduces `PartitionFunction` as an interface. Partitioning function
is a general concept useful in different places (remote or local
exchange partitioning or partitioning spiller for join's spill) so it
deserves simple expressive name.
